+++ elf_eh.h
@@ -36,5 +36,5 @@ struct eh_fde_rel {
 struct ElfEHInfo {
   bool MeasureFrame(const eh_frame_hdr* hdr,
                     uintptr_t* eh_frame_ptr,
-                    size_t* eh_frame_len);
+                    uint32_t* eh_frame_len);
 };
